+++ src/main/java/com/team2502/ezauton/localization/sensors/ITachometer.java
@@ -4,5 +4,5 @@ public interface ITachometer extends ISensor {
     /**
      * @return revolutions / s
      */
-    float getVelocity();
+    double getVelocity();
 }
